Makes AdaptiveSampleRate usable in java
This reorganizes code so that the AdaptiveSampleRate class is usable in
Java. This allows those users to begin testing with compatible sampling.
It also allows us to focus energy, by eliminating the overhead of
maintaining https://github.com/openzipkin/zipkin-sampler-adaptive
