+++ zipkin-sampler/src/main/scala/com/twitter/zipkin/sampler/ZKClient.scala
@@ -16,6 +16,10 @@
 
 package com.twitter.zipkin.sampler
 
+import java.net.InetSocketAddress
+import java.util.concurrent.Executors
+import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
+
 import com.google.common.base.Supplier
 import com.twitter.common.base.ExceptionalCommand
 import com.twitter.common.quantity.{Amount, Time => CommonTime}
@@ -27,37 +31,32 @@ import com.twitter.finagle.util.DefaultTimer
 import com.twitter.logging.Logger
 import com.twitter.util._
 import com.twitter.zk.{Connector, ZNode, ZkClient}
-import java.net.InetSocketAddress
-import java.util.concurrent.Executors
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
 import org.apache.zookeeper.{CreateMode, KeeperException, Watcher, ZooDefs}
+
 import scala.collection.JavaConverters._
 import scala.collection.mutable
 
-trait ZkWatch[T] extends Closable {
+private[sampler] trait ZkWatch[T] extends Closable {
   val data: Var[T]
 }
 
-object ZKClient {
+private[sampler] object ZKClient {
   val pool: FuturePool = new ExecutorServiceFuturePool(Executors.newCachedThreadPool(
     new NamedPoolThreadFactory("ZKClientPool", makeDaemons = true)))
 }
 
 // TODO: rewrite with curator recipes or similar
-class ZKClient(
-  addrs: Seq[InetSocketAddress],
-  credentials: Option[(String, String)] = None,
+private[sampler] class ZKClient(
+  servers: java.util.List[InetSocketAddress],
+  credentials: ZooKeeperClient.Credentials,
   timeout: Duration = 3.seconds,
   stats: StatsReceiver = DefaultStatsReceiver.scope("ZKClient"),
   log: Logger = Logger.get("ZKClient"),
   timer: Timer = DefaultTimer.twitter,
   pool: FuturePool = ZKClient.pool
 ) extends Closable {
-  private[this] val client = credentials map { case (u, p) =>
-    new ZooKeeperClient(Amount.of(timeout.inMilliseconds.toInt, CommonTime.MILLISECONDS), ZooKeeperClient.digestCredentials(u, p), addrs.asJava)
-  } getOrElse {
-    new ZooKeeperClient(Amount.of(timeout.inMilliseconds.toInt, CommonTime.MILLISECONDS), addrs.asJava)
-  }
+  private[this] val client =
+    new ZooKeeperClient(Amount.of(timeout.inMilliseconds.toInt, CommonTime.MILLISECONDS), credentials, servers)
 
   private[this] val zkClient = ZkClient(new Connector {
     client.register(sessionBroker)
