+++ zipkin-sampler/src/test/scala/com/twitter/zipkin/sampler/StoreRateCheckTest.scala
@@ -46,7 +46,7 @@ class StoreRateCheckTest extends FunSuite {
   test("allows only once per period") {
     Time.withCurrentTimeFrozen { tc =>
       val timer = new MockTimer
-      val check = new CooldownCheck[Unit](1.minute, NullStatsReceiver, timer = timer)
+      val check = new CooldownCheck[Unit](60, NullStatsReceiver, timer = timer)
 
       assert(check(Some(())).isDefined)
       assert(check(Some(())).isEmpty)
@@ -83,7 +83,7 @@ class StoreRateCheckTest extends FunSuite {
 
   test("calculates a discounted average based on the current req rate and sample rate") {
     val tgtStoreRate = Var(100)
-    val sampleRate = Var(1.0)
+    val sampleRate = Var(1.0f)
     val calc = new CalculateSampleRate(tgtStoreRate, sampleRate, DiscountedAverage, 0.05, 1.0)
     assert(calc(Some(Seq(1000, 1000, 1000))) === Some(0.1))
   }
